+++ packages/moonstone/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/Popup` to properly handle closing in mid-transition
+- `moonstone/Picker` not to read out previous or next item when the decrementer or incrementer icon is disabled
 
 ## [3.0.0-alpha.7] - 2019-06-24